# Conflicts:
#	Common/ac/messageinfo.cpp
#	Common/ac/point.cpp
#	Common/ac/roomstruct.cpp
#	Common/ac/roomstruct.h
#	Common/ac/wordsdictionary.cpp
#	Common/util/compress.cpp
#	Common/util/compress.h
#	Editor/AGS.Native/agsnative.cpp
#	Engine/ac/gamestate.h
#	Engine/ac/global_drawingsurface.cpp
#	Engine/plugin/agsplugin.h
#	Engine/script/script_engine.cpp
#	Solutions/Common.Lib/Common.Lib.vcxproj.filters

+++ Common/util/textreader.h
@@ -35,11 +35,11 @@ public:
     // Read single character
     virtual char    ReadChar()              = 0;
     // Read defined number of characters
-    virtual String ReadString(int length)  = 0;
+    virtual String  ReadString(size_t length) = 0;
     // Read till line break
-    virtual String ReadLine()              = 0;
+    virtual String  ReadLine()              = 0;
     // Read till end of available data
-    virtual String ReadAll()               = 0;
+    virtual String  ReadAll()               = 0;
 };
 
 } // namespace Common
